+++ src/pxlNav/effects/particles/HeightMap.js
@@ -2,21 +2,27 @@
 import {
   Vector3,
   NearestFilter,
+  LinearFilter,
   NearestMipmapNearestFilter,
+  LinearMipmapLinearFilter,
   AdditiveBlending
 } from "../../../libs/three/three.module.min.js";
 
 import ParticleBase from './ParticleBase.js';
-import { dustVert, dustFrag } from './shaders/FloatingDust.js';
+import { heightMapVert, heightMapFrag } from './shaders/HeightMap.js';
 
 // Free floaties in the environment
 //   Dust balls & flakes
 
-export class FloatingDust extends ParticleBase{
-  constructor( room=null, systemName='floatingDust'){
+export class HeightMap extends ParticleBase{
+  constructor( room=null, systemName='heightMap'){
     super( room, systemName );
     this.name=systemName;
     this.room=room;
+
+    this.heightMapPath = null;
+    this.spawnMapPath = null;
+    this.spawnMapMode = 1;
 
     this.material = null;
 
@@ -30,7 +36,6 @@ export class FloatingDust extends ParticleBase{
       "randomAtlas" : false,
       "additiveBlend" : false,
 
-      "windDir" : new Vector3( 0, 0, 1 ),
       "offsetPos" : new Vector3( 0, 0, 0 ),
 
       "hasLights" : false,
@@ -42,9 +47,23 @@ export class FloatingDust extends ParticleBase{
     this.knownKeys = Object.keys( this.shaderSettings );
   }
   
+  setHeightMapPath( path ){
+    this.heightMapPath = path;
+  }
+
+  setSpawnMapPath( path, channels=1 ){
+    this.spawnMapPath = path;
+
+    // If the spawnMap includes wind direction data in Green & Blue channels
+    if( channels == 3 ){
+      channels = 4;
+    }
+    this.spawnMapMode = channels;
+  }
+
   // 'vertexCount' - Point Count
   // 'pScale' - Point Base Scale
-  build( curShaderSettings={} ){
+  build( curShaderSettings={}, objectRef=null ){
     
     if( curShaderSettings && typeof curShaderSettings === Object ){
       let curSettingKeys = Object.keys( curShaderSettings );
@@ -56,9 +75,9 @@ export class FloatingDust extends ParticleBase{
         }
       });
     }
-
-
-    if( !this.shaderSettings["atlasPicks"] || this.shaderSettings["atlasPicks"].length < 1 ){
+    if( curShaderSettings.hasOwnProperty("atlasPicks") ){
+      this.shaderSettings["atlasPicks"] = curShaderSettings["atlasPicks"];
+    }else if( !this.shaderSettings["atlasPicks"] || this.shaderSettings["atlasPicks"].length < 1 ){
       this.shaderSettings["atlasPicks"] = [
         ...super.dupeArray([0.0,0.],4), ...super.dupeArray([0.25,0.],4),
         ...super.dupeArray([0.5,0.0],2), ...super.dupeArray([0.5,0.25],2),
@@ -72,12 +91,6 @@ export class FloatingDust extends ParticleBase{
 
     // -- -- --
 
-    // Clean up any {} settings that should be Vector3s
-    let windDir = curShaderSettings["windDir"];
-    if( windDir && typeof windDir === Object && typeof windDir !== Vector3 && windDir.length === 3 ){
-      this.shaderSettings["windDir"].set( windDir[0], windDir[1], windDir[2] );
-    }
-
     let offsetPos = curShaderSettings["offsetPos"];
     if( offsetPos && typeof offsetPos === Object && typeof offsetPos !== Vector3 && offsetPos.length === 3 ){
       this.shaderSettings["offsetPos"].set( offsetPos[0], offsetPos[1], offsetPos[2] );
@@ -85,7 +98,31 @@ export class FloatingDust extends ParticleBase{
 
     // -- -- --
 
+    let sizeX = 100;
+    let sizeY = 100;
+    let sizeZ = 100;
+    if( objectRef && objectRef.hasOwnProperty("userData") ){
+      if( objectRef.userData.hasOwnProperty("SizeX") ){
+        sizeX = objectRef.userData.SizeX;
+      }
+      if( objectRef.userData.hasOwnProperty("SizeY") ){
+        sizeY = objectRef.userData.SizeY;
+      }
+      if( objectRef.userData.hasOwnProperty("SizeZ") ){
+        sizeZ = objectRef.userData.SizeZ;
+      }
+    }
+    let tankSize = new Vector3( sizeX, sizeY, sizeZ );
+
+
+
+    // -- -- --
+
     let dustUniforms={
+      heightMap:{ type:"t", value: null },
+      spawnMap:{ type:"t", value: null },
+      tankSize:{ type:"v", value: tankSize },
+
       atlasTexture:{ type:"t", value: null },
       atlasAlphaTexture:{type:"t", value: null },
       noiseTexture:{ type:"t", value: null },
@@ -93,13 +130,12 @@ export class FloatingDust extends ParticleBase{
       pointScale:{ type:"f", value: this.pscale },
       intensity:{ type:"f", value:1.0 },
       rate:{ type:"f", value:.035 },
-      positionOffset:{ type:"v", value:this.shaderSettings["offsetPos"] },
-      windDir:{ type:"v", value:this.shaderSettings["windDir"] }
+      positionOffset:{ type:"v", value:this.shaderSettings["offsetPos"] }
     };
     //let mtl = this.pxlFile.pxlShaderBuilder( snowUniforms, snowFallVert( true ), snowFallVert() );
 
     
-    let mtl = this.room.pxlFile.pxlShaderBuilder( dustUniforms, dustVert( this.shaderSettings ), dustFrag( this.hasAlphaMap ) );
+    let mtl = this.room.pxlFile.pxlShaderBuilder( dustUniforms, heightMapVert( this.shaderSettings ), heightMapFrag( this.hasAlphaMap ) );
     mtl.transparent=true;
 
     if( this.hasAlphaMap ){
@@ -109,6 +145,14 @@ export class FloatingDust extends ParticleBase{
       }
     }else{
       mtl.uniforms.atlasTexture.value = this.room.pxlUtils.loadTexture( this.atlasAlphaPath, 4, {"magFilter":NearestFilter, "minFilter":NearestMipmapNearestFilter} );
+    }
+    
+    if( this.heightMapPath ){ // RGB Height Map
+      mtl.uniforms.heightMap.value = this.room.pxlUtils.loadTexture( this.heightMapPath, 4, {"magFilter":LinearFilter, "minFilter":LinearMipmapLinearFilter} );
+    }
+    
+    if( this.spawnMapPath ){ // A or RGB Spawn Map
+      mtl.uniforms.spawnMap.value = this.room.pxlUtils.loadTexture( this.spawnMapPath, this.spawnMapMode, {"magFilter":NearestFilter, "minFilter":NearestMipmapNearestFilter} );
     }
 
     if( this.shaderSettings["additiveBlend"] ){
@@ -125,6 +169,10 @@ export class FloatingDust extends ParticleBase{
     this.material = mtl;
 
 
-    super.addToScene();
+    let pSystem = super.addToScene();
+    pSystem.position.set( objectRef.position.x, objectRef.position.y, objectRef.position.z );
+    pSystem.userData["tankRes"] = tankSize;
+
+    return pSystem;
   }
 }