+++ src/pxlNav/effects/particles/shaders.js
@@ -4,8 +4,9 @@
 //import { countDownVert, countDownFrag  } from './CountDownShader.js';
 import { emberWispsVert, emberWispsFrag } from './shaders/EmberWisps.js';
 import { dustVert, dustFrag } from './shaders/FloatingDust.js';
+import { heightMapVert, heightMapFrag } from './shaders/HeightMap.js';
 import { smokeVert, smokeFrag } from './shaders/Smoke.js';
 //import { snowConfettiVert, snowConfettiFrag  } from './SnowConfettiShader.js';
 import { snowFallVert, snowFallFrag } from './shaders/SnowShader.js';
 
-export { emberWispsVert, emberWispsFrag, dustVert, dustFrag, smokeVert, smokeFrag, snowFallVert, snowFallFrag };
+export { emberWispsVert, emberWispsFrag, dustVert, dustFrag, heightMapVert, heightMapFrag, smokeVert, smokeFrag, snowFallVert, snowFallFrag };