+++ src/pxlRooms/TemplateEnvironment/TemplateEnvironment.js
@@ -1,6 +1,6 @@
 
 import * as THREE from "../../three.module.min.js";
-import { pxlEffects, RoomEnvironment } from "../../pxlNav.esm.js";
+import { pxlEffects, RoomEnvironment } from "../../pxlNav.js";
 
 // Lets add a particle effect to the room
 //   This is lingering dust that floats around the user to give a sense of movement